Exclude if `onFramework` is None
My initial assumption was that the presence of a supplier_framework
record indicated that supplier had applied to that framework.
In reality, all suppliers with a (partially) completed declaration have
a supplier_framework record, whether or not they correctly applied (ie,
they need also to have submitted G7 services).
Solution was to exclude supplier_framework records where the value for
`onFramework` was `None`.