Check that frameworks parameter is a string
If the command-line flag `--frameworks=` exists with no value then it is treated as a boolean option and given the value True.  In this case no frameworks would be indexed because `True` would be passed through as the `framework` parameter to the call to get services.  Instead we want all live services to be indexed if no frameworks are specified, which is what happens if we set the value to `None` instead of `True`.