Fix iter_services not raising an exception when used with Pool.imap
Multiprocessing `Pool.imap` doesn't raise an exception if the iterator
fails to produce any results (ie when the first API request fails),
making the process exit with a 0 status. This will mark the CI job
as successful when the indexing failed to run.

Replacing `Pool` with `ThreadPool` seems to fix the problem, and other
scripts are already using `ThreadPool`. Also replacing `.imap` with
`.imap_unordered` since we don't really care about the response order.