Evaluate DOS applications for pass/fail
For each supplier interested in DOS the script checks whether:
 - They submitted a complete declaration with all the right answers
 - The lot-specific essential questions were answered correctly
 - The supplier has some valid services submitted after failed lot applications are discounted)

And decides whether the application is a PASS, FAIL or DISCRETIONARY award.
For PASS the `supplier_frameworks` entry for `on_framework` is set to True.
For FAIL the `supplier_frameworks` entry for `on_framework` is set to Flase.
For DISCRETIONARY the `supplier_frameworks` entry for `on_framework` is left as Null.