Export DOS applications
Export DOS application information into three CSV files for successful,
failed and discretionary applications.

This script assumes that `insert-dos-framework-results.py` has already
been run and will fail with an assertion error if it has not (when it
comes across a declaration that is not complete while the supplier is
not marked as failed.

The file of successful applicants contains supplier information, contact
details and counts for completed, failed and draft services.
The file of failed applicants contains supplier information, a list of
failed mandatory questions, contact details and counts for completed,
failed and draft services.
The file of discretionary applicants contains supplier information,
discretionary questions and their answers, contact details and counts
for completed, failed and draft services.

Service level failed questions are not included in these reports.