Add script to generate export data CSV files for CCS
This script and the associated tests are based on the existing `export_dos_suppliers` script.
This version is somewhat simpler because there are no lot-level declaration questions for G-Cloud, and so no concept of a "failed" service.
Also, the data need of CCS has been revised, so there is less data included in these files (e.g. no service counts and less "about me" type data).