Catch errors on individual services to allow script to continue
If there's a connection error or something similar then we don't want the whole script to fail. Wrapping the `make_draft_service_live` code in a try...except allows the script to continue running, and reports the error so that individual failures can be corrected at the end of the run.