Explicitly return supplier ids as a list
In python3, supplier_ids were being returned as filter objects
which can only be iterated through once.
This meant that once we got to the end of the iteration, we would
stop matching suppliers at all and not end up resetting all the
framework agreements we wanted (but we would reset some of them).

Wrapping the filter in a list fixes behaviour in python 3 and
preserves the correct behaviour in python 2.