Return all table row data and print to CSV
We were looking for a more regular way of giving data to our
data analysis team.
The solution was to come up with a way of printing out table
rows and format it in a CSV file.
Using these scripts, we can:
1. get all rows for any model with an `iter` method
2. pick a set of keys we would like to turn into columns
  - including nested keys (suppliers have `contactInformation`, for example)
3. filter the resulting list based on the value of a key
  - eg, a user's `role` must equal `supplier`
4. do some additional processing on returned values
  - eg, strip the username from an email address
5. sort the list based on the values of a key
  - eg, sort users based on the date they signed up
6. print everything to a csv

In the future, we are planning to add logic to JOIN data from one
table to another it references, but for now this is what we have.