Add script to generate counterpart G-Cloud 8 PDFs
I started off parameterising the `generate-agreement-sinature-pages` script, but it turns out that there are enough differences with generating the counterpart pages that it makes more sense to me as a separate (but very similar) thing.

The main differences are different conditions on when a file should be generated, an extra file that needs to be copied and later removed, (i.e. the signature file), and file names.

It could probably have been parameterised and just one script, but would just involve an explosion of command line arguments that wouldn't be very pretty.