Add script to upload countersigned agreement files
This will upload a folder of counterpart pdf files to S3 and mark the countersignedAgreementPath in the API for the relevant FrameworkAgreement.

Files are uploaded to S3 with a unique (timestamped) filename, and the Content-Disposition header is set to a more user-friendly filename that includes the supplier name and without a timestamp.