Add export-framework-suppliers script
This is a version of the existing export_dos_suppliers and export_g8_suppliers script
that should hopefully be general purpose for new frameworks going forward.

Rather than explicitly encoding the "right" answers to declaration questions in the
script this version uses a JSON schema (supplied in the arguments) to determine
which questions caused a "fail" or "discretionary" result.