Join CSV with group counts
Adds additional operations support to the get-model-data scripti config.

CONFIG is now an ordered list of items. Each item has a name and can be
requested or skipped when the script is invoked.

Since items are processed in order, each one can depend on CSVs for the
previous ones existing in the OUTPUT_DIR.

Each CONFIG item can describe configuration for several steps, which are
invoked by the script in order. First step is loading the data from one
of the sources:

* Data API, if `base_model` key is set
* Existing CSV file if `model` key is set
* A join of two existing CSVs if `join` key is set

Once the data is loaded it is transformed into pandas `DataFrame`
object, which is then used by the other processing steps.

Currently supported config keys include:
1. `process_fields` defines individual field transformations to be
   applied to the data
2. `add_counts` adds count columns for groups of related data (eg
   counts of brief responses grouped by status for each brief)
3. `filter_query` is a pandas query expression to filter output
   CSV rows
4. `keys` is setting the set and order of output CSV columns
5. `sort_by` sets the list of columns to use for ordering the CSV