Remove unnecessary loop
The call to `send_email` includes all users for the brief, so no need to loop through them here (in fact, if there were more than one user for a brief then everyone would receive multiple emails).

I've left the test that the brief actually has some users to email in as a "just in case" because we did have some orphaned briefs and don't want to attempt a send to no email addresses for them.