Remove one-off scripts to insert G8 and DOS results
These scripts were used to assess the automatic passes and fails for G-Cloud 8 and DOS frameworks.  In future we are not going to do this the same way, instead moving to the JSON schema-based way we did for DOS2 - see  "mark-definite-framework-results.py".

No need to keep these in the codebase now.