Add `count_statuses=("submitted",)` to call to get CSV rows
Technically this additional parameter isn't required for G-Cloud, as G-Cloud doesn't have a concept of failed services. But if it did have the concept of failed services then we'd want to exclude them from the lots awarded in the framework agreement, and adding this parameter would ensure that that happened.