Fix logic for including suppliers in applicant details script
Previously the condition only checked for the existence of `record['counts']`, which meant that rows were being generated for suppliers who did not have a service with any of the desired statues - e.g. suppliers with a complete declaration but only draft services were included in the CSV output from `export-framework-applicant-details` but with zero service-counts for all lots.  This fixes things so `_create_row` is only called for records with services in the right status (e.g. only suppliers who submitted a service).