Fix python 2/3 IOError message attribute difference
`RequestException` inherits from `IOError`. In Python 2, `IOError`
has a `message` attribute so you can call `e.message`. However in
Python 3 `IOError` is an alias for `OSError` which does not have a
`message` attribute`. This change is poorly (not) documented online.

To solve this issue we instead use `str(e)` rather than `e.message`.