Fix mailchimp capitilisation bug
We believe if an email is added to mailchimp for the very first time
(maybe not even by ourselves) then mailchimp stores both the
email address (which could contain capitals) and then a hash of the
email lowercased. Then when we query for all emails on our list
it returns the email addresses for all the hashes which are on our
list.

This means we could have a lowercase version of an email in our API
but mailchimp has a capitalised version stored. Mailchimp however
treats these emails as the same because it uses it's hash to
identify.

Lowering our lists of emails means we return the correct number
of emails which have not yet been added to a mailchimp list.