Add log statement for non-dry run
Clarification Questions are not returned by the API by default.
Changes are required there to allow an optional param to return
this data. In the meantime, the script will pass without it (but
won't send any emails, as no questions have been included in the
query). (API changes have now been made)