Add an optional argument to exclude supplier ids
This is done in case the script breaks midway through execution.
`main()` filters out the excluded IDs.

On the way we encountered an issue with docopt: when breaking the
line for example use, do not start new line with the dash or docopt
might think you are defining the same argument twice.