Catch email failures and log supplier ids to be passed into script
We change our script from taking a list of supplier IDs to exclude
to list of supplier IDs to restrict to. This is supported
because we now catch any email errors and log a list of
all suppliers that we could not send an email to. This
list can then nicely be passed into the script so it can
run for a second attempt without emailing anyone twice. This
solution makes it much easier for a developer to run the
script after errors have occured without needing to manually
parse supplier IDs from several log messages.