Add script to programatically update an ES alias
Currently adding or moving an elasticsearch index alias is a manual job
involving constructing a bespoke request with curl.

You can do the same thing by giving this script the arguments required.
It will fetch the correct auth token for the search api from the
credentials repo.

It's being written for use by Jenkins, but could be used locally.

If invoked with `--delete-old-indexes-yes` it will remove any indexes
from the given stage which do not have an alias.