Set --delete-old-indexes default to 'no'
By having the default as nothing docopt would assign it an empty string.
This was useful if the check later was for a 'truthy' or 'falsey' value.
However, if you set the value to ANY string ('no' for example), it
becomes a 'truthy' value which isn't helpful.

Because of this the check is for a string literal, 'yes'. This means we
can set the default to 'no'.