Rename current alias to old and don't delete index
By renaming the alias on the index that is losing the alias, instead of
just removing it, we can keep that index around. This means that if
something goes wrong, we have the last index to fall back to quickly.

This commit also refactors the response check to avoid duplication of
code.