Only delete the index which is losing the `old-<alias>` alias.
It was felt that deleting all indexes with no alias was a bit too
strong.

This updates to only delete the index which is losing the `old-<alias>`
alias.

Also, use strtobool to convert the value passed to `--delete-old-index`
to a 1 or 0, which can then be treated as a boolean.