Add freeze-requirements
At the moment, scripts running in Jenkins are using the local
venv created by the Jenkins job, which means they won't update
any packages unless a version is changed explicitly.

This means that recreating that venv could end up with a different
set of package versions than what's currently installed.

We need to freeze our nested package requirements to make sure
we're not accidentally updating packages between builds.

Similar to the application changes, this renames requirements.txt
to requirements-app.txt. Updates Makefile to add freeze-requirements.

requirements-app.txt is not a great name for the scripts repo, but it's
consistent with what we call it in other repos.

This also switches Travis to use `make test` target instead of running
py.test directly. `make requirements` is setting up dependencies in
the virtualenv, but running `py.test` directly isn't activating that
virtualenv explicitly. The only reason this worked on Travis is because
it's activating a clean virtualenv before running any steps for the
python builds.

NOTE: This is upgrading dmutils to bring a newer version of the cryptography
package in Python 3. There's a major change related to S3 in the new
utils version, so scripts using it might need to be updated.