get-model-data: fix missing group columns error
When filtering output csv keys the script expects all column to
exist. But additional "grouped" columns might be missing from the
table if there were no values for the column.

For example, our preview and staging environments don't have any
brief responses with failed essential requirements, so the column
is never created and the field filter raises an exception.