Log list of supplier IDs that we intend to send emails to
If our script fails middway, we rerun it and include a specific
list of supplier IDs only to send to.  If the script fails middway
though we don't know which IDs these are as they are never logged
out. This commit will log them meaning they can be easily be
copied and pasted back into the job (from the supplier ID of
failure).