Retry more errors and skip already-done
The script stopped a couple of times trying to run it so add this to retry on more "permanent" looking errors that in reality can be recovered with a single retry.

Also don't keep posting updates for the same suppliers, as this makes it take really long time to run.