Add `get_model_by_fk` query
Sometimes you want to get data from the api that requires the id of a
different model to be supplied in the request. If you want to get that
data for a bunch of different model id's.

An example of this is direct_award_project_searches. The url on the api
to get them is `/direct-award/projects/<int:project_id>/searches`.

If we want to retrieve all the searches from the api we first need a
list of all the projects, and the iterate over their id's.

It's also useful to be able to filter the original models before
fetching the data via their id's.

It's also useful to be able to group and count them before returning
the data, in case it doesn't have a key to group by later.