Handle empty data gracefully
If data was requested but none returned due to an empty table for
example, the get-model-data scripts would often fall over due to
key-errors when trying to perform some action later on.

This is due to the queries just returning an empty data frame with no
column headings.

By returning a DataFrame with the expected column headings, we can
mitigate this.