Revert changes to ModelTrawler
The ModelTrawler was changed to be able to use `get_<model>_iter`
methods in the api client as well as `find_<model>_iter` methods. This
was for the sake of one method,
`get_direct_award_project_services_iter`, which is the only `get_`
method.

The easier solution is to set up an alias in the apiclient`, like so:

`find_direct_award_project_services_iter` =
`get_direct_award_project_services_iter`

This way, the apps can still use the `get_` method and the script can
use the `find_` method.