Add one-off script for renaming registration country
Currently suppliers that are registered in the UK have the string 'gb'
set in the registration_country field of the suppliers table.

As part of the new supplier sign up flow we're going to be collecting
their registration country via the country picker supplied by registers.

We need to bring our existing data in line with the data the country
picker will be supplying.