Migrate supplier trading status
One-off script for migrating suppliers existing trading status
(submitted during framework applications) to our new set of values now
that this is an account-level question. If we can map it perfectly
across, we'll do so, but if we can't then we'll drop the value
altogether - the result of this being the supplier will need to provide
it again.