Create maintenance mode PR
 ## Summary
Adds a script to create a Pull Request on the dm-aws repository which
toggles maintenance mode on/off for a specific environment, in the same
vein as the script for rotating API tokens. Not ready to generalise this
script yet - but if we add another job that manages Pull Requests we
should review and see what can be taken out and centralised.