Fix bug with org size migration script
The script was ordering supplier frameworks by there
`agreementReturnedAt` date. In some instances however the agreement
hadn't been returned and so this value was None. These instances seemed
to be either when a supplier was not on framework and so potentially had
a half complete declaration, or some G7 supplier frameworks didn't have
this date set.

When ordering by `agreementReturnedAt`, any framework with `None` for
this date would always appear at the beginning of the list and so be
selected. The knock on effect being that there was a high chance they
it didn't have a correct value for `organisationSize` in the declaration
and so the suppliers organisation_size would not be set.

This occurred even if a supplier had other supplier_frameworks which
could have been used successfully.

By dropping supplier frameworks that don't have `agreementReturnedAt` we
can side step this issue.