Update notify suppliers script
 ## Summary
We don't need to pass the `successful notification date` in any more as
this is now stored on the API. It is called the `intention to award at`
date there, so we'll update references to keep language consistent.

I'll update the two Notify templates using this personalisation when
this PR is merged.