Update generate framework master csv
 ## Summary
We now inject company details into the supplier framework when a
framework changes from open to pending. This means that suppliers who
haven't started a declaration no longer have a 'False-y' declaration if
they haven't answered any questions, so we need to do a `get('status')`
rather than a direct `['status']` on the dict.