Add an entrypoint to forward API ports from Docker
This is useful for macOS development environments, if services are
running on the host machine they are not normally accessible to the
container.

This commit adds a script `docker_for_mac_entrypoint.sh`, that
forwards the ports within the container transparently to the Docker
host, so that the API services can be accessed as if they were
running within the container.

By default the entrypoint is not used; it is copied into PATH in the
container but must be enabled when `docker run` is called. An
explanation of how to do this has been added to the readme file.