Redirect Notify emails in scripts
 ## Summary
We send a lot of fake emails with scripts that run against preview.
Doing this regularly and repeatedly will degrade Notify's reputation
with its email provider, so we need to stop sending emails to addresses
that we know are bad. We should make a concerted effort to stop doing
this wherever possible.

This is an alternative to centralising the email redirects in utils as
put forward in
https://github.com/alphagov/digitalmarketplace-utils/pull/405.

 ## Ticket
https://trello.com/c/YIJ25eiP/479