Generate framework agreement pages
 ## Summary
We have a separate `digitalmarketplace-agreements` repository that holds
templates used to generate the signature page for framework agreements
that successful suppliers for G-Cloud and Digital Outcomes and
Specialists need to sign before providing services on the framework.
This repository used to have a separate template for each framework
iteration, but now we have a unified template for the signature page
that all frameworks (on the Digital Marketplace, at least) can use.
This repository is pointed at by a script in this
(digitalmarketplace-scripts) repository, so it introduces a layer of
separation between two related things. Given that we are only
maintaining a single template now, effectively, it feels better to bring
the template into the scripts repository directly to minimise the number
of repos we need to juggle and to bring the two related things closer
together so that it's easier to find and maintain context.

This commit effectively deprecates the digitalmarketplace-agrements
repository. It creates a new top-level folder here, called 'templates',
which contains a 'framework_signature_page' to hold the template. A
separate top-level folder, 'email_templates', has been consolidated
under the new 'templates' top-level directory.

The framework agreement template has been updated slightly to update
variable names based on how we now store company details in the supplier
declaration, and how we have removed some duplicate keys from the
framework.frameworkAgreementDetails data blob.