Agreements per framework
 ## Summary
We still theoretically may need to countersign DOS2 framework
agreements. The signature page has changed for G10, so we need to have
separate templates for each of these frameworks again. This adds a DOS2
framework and includes the necessary legacy foramtters (mostly around
datetime) to generate countersigned pages that look the same as they do
at the moment.

Because we still need to maintain a separate agreement template for
DOS2, we're going to reinstate the `digitalmarketplace-agreements`
repository. There was a dream that we could have a single signature
page for all framework agreements, but this seems to not be a reality
at the moment, so that repository remains a sensible place to store the
templates. Therefore we're removing the signature page templates from
this repository. The countersignature image will be stored next to the
template, as it was before.