Convert country codes to names
 ## Summary
We store supplier country details in the database as country codes that
relate to a Country or Territory Register. These are not intended for
human consumption, so we should convert these back to the proper
representation of the country (e.g. `country:GB` -> `United Kingdom`),
otherwise we need to do this manually before sharing the data with our
partners (e.g. CCS).

 ## Ticket
https://trello.com/c/bx6MXS0V/173