Update list of keys to check in bad words script
The list is now an amalgamation of all keys checked for G7 through G10.

I did want to remove this list completely and just scan any value in the
service data that was of string type. Unfortunately we have list
questions, and checkbox questions, which both appears as lists of
strings in the data.

The strings from checkbox questions are defined by us, and they may
include words defined in our list of bad words (words given to us by
CCS, not naught words). This means we could get some false positives.