Acknowledge draft service updates
 ## Summary
When moving G-Cloud 10 services from drafts to 'published', we generate
a lot of `update_service` audit events. We have an admin page that
allows CCS to review and acknowledge `update_service` audit events, but
this is only intended for services which are currently live (i.e. the
framework is open). Service changes from the script which promotes the
drafts, `publish-g-cloud-draft-services.py` (and
`publish-dos-draft-services.py`) should not need to be acknowledged by
CCS.

This script will go through and acknowledge all audit events created by
the publishing scripts.