Add script for successful suppliers csv
 ## Summary
When a framework goes live we need to produce a CSV file of all the
suppliers that were successfully awarded onto the framework and all the
lots they were successful on. This is then uploaded to the frameworks
communications updates page. At the moment I guess this was done
manually, as I can't see a script that produces this file.

So now we have one.

 ## Ticket
https://trello.com/c/aGxhSBwa/150