Remove code that relies on sops/credentials repo
When running scripts in Docker, the credentials repo and SOPS are not 
available.

If the user forgets to expose the api token environment variable, then 
they will get confusing errors about SOPS not being available, when 
really the error should be that the api token is not available in the 
environment.

As we are going to be running scripts in Docker from now on, this code 
is dead, and so I have replaced it with a helpful error message.