Prevent github api token from being echoed in rotate api tokens script
rotate-api-tokens.sh uses a GitHub API token to get the
digitalmarketplace-credentials repo; however when a push is made the
token is echoed to the screen, which means it will end up in all sorts
of logs.

This commit adds the quiet flag to the push so that it will only echo if
there is an issue.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Katherine Stevens <katherine.stevens@digital.cabinet-office.gov.uk>