Remove hard coded framework from send_dos_opportunities
The script to send suppliers information about newly released dos
opportunities had the framework name hardcoded into the subject line for
the mailchimp campaign. This was an issue as there is a chance that
we'll have new briefs from two different iterations when we transition
from DOS2 to DOS3.

This grabs the iteration from the live briefs found and formats the
subject line. It also means we don't have to worry about updating the
script after DOS3 goes live.