threading: make all uses of multithreaded map do so through dependency injection
this way different implementations can be supplied on invocation and e.g.
we can call each in single threaded mode by supplying the builtin `map`,
which is also used as the default argument everywhere so these functions
can still be called from anywhere without worrying about the map
implementation and still get a safe result.

this is particularly useful for tests, where running the code in
multithreaded mode complicates things and can even cause deadlocks if an
instance of `mock.Mock` is shared among several threads.