data_retention_remove_user_data: tweak to make less likely we shoot selves in foot
because succeeding in removing a users personal data from the API will
cause their email address to be scrubbed, if we make removing the data
from the API the *first* thing we attempt to do, we will only give
ourselves one chance to remove the address from mailchimp, because a
subsequent run of the script won't have access to the email address in
question anymore.

simple fix is to attempt the mailchimp removal first and ensure we don't
then continue on to remove a user from the API db if we haven't been able
to remove them from mailchimp.

the loud failure should spur one of us to do something about it or
reconsider how we should handle this.