Add basic script for generating categories/reseller report
This adds a script requested by our performance analyst. It gives them data about which of our suppliers are resellers and which are selling original products. It also lists the categories each service has registered under. The script should be run against a local dump, as it hammers the API fairly hard. It outputs to TSV, because working with commas is hard and this was the path of least resistance.