Fix csv export so only aborts if no data found
We had an issue where CSVs with no rows were being created legitimately
(because there were no entries in the database that fit the criteria),
but the CSV export script was failing to upload this to s3 as it thought
this was an anomaly.

This commit changes the `generate_csv_and_upload_to_s3` procedure so it
only aborts if it does not have a valid s3 filename.