Move embedded IPython shell into the global namespace
Calling `IPython.embed()` in a function means the shell will be created
in the scope of that function, which has a bunch of annoying
side-effects, such as function definitions not being available in other
closures.

The workaround seems to be to instead use `IPython.start_ipython()`, and
to do so in the global namespace (see [1]).

[1]: https://ipython.readthedocs.io/en/stable/interactive/reference.html#embedding-ipython