Change the default port for the search api
The GDS-managed macOS system includes a VPN tool called Cisco
AnyConnect; the latest version of this has a management daemon that
starts at boot and listens on port 5001, conflicting with the port for
the search api.

As we are unable to change the VPN tool, instead we have to change the
port for the search api :(

This commit updates digitalmarketplace-utils from 48.1.0 to 49.0.0,
pulling in the changes from alphagov/digitalmarketplace-utils#540, which
changes the environment helpers we use to find the URLs for each
environment.