Add brief-status option to get-user script
As a team member testing the Digital Marketplace...
I want to be able to find a user with DOS opportunities in a certain state...
So that I can check different parts of the DOS journey.

This commit adds a command line flag `--brief-status=<status>` that
tries to make it easier to do this. The UI is a bit unhelpful and you
have to look at the api[1] to have an idea of what the different
statuses mean, but it's a start?

[1]: https://github.com/alphagov/digitalmarketplace-api/blob/release-1029/app/models/main.py#L1543-L1591