Remove notify-suppliers-of-new-fw-cq-answers script
We have two different scripts that can send email 21 "New framework
clarification questions and answers have been uploaded". This is because
we intended to replace the script notify-suppliers-of-fw-cq-answers with
the more generic and capable script notify -suppliers-of-framework-application
-event, but we never got around to it (see
alphagov/digitalmarketplace-scripts#475 for the history of this).

However, having the two different scripts has caused some confusion, and
we've been happily using notify-suppliers-of-framework-application-event
for some time, so we can get rid of the older script.