Allow read-only access to the API client shell
Having direct access to production faintly terrifies me. Sometimes I just need to get information from prod. In those cases, it'd be useful to have protection against accidentally making changes.

Override __getattr__ for access control. Override __dir__ so that autocomplete in the iPython shell works.

I've done this on the assumption that API methods that start with "get_" or "find_" are read-only methods.