Fix bug in upload counterpart agreements script
Previously the upload counterpart agreements script would assume that
the documents it was uploading were for the suppliers latest framework
agreement. This could lead to a race condition where a counterpart
agreement would be uploaded to an agreement that had not been approved.

This commit updates the agreement with the ID in the agreement document
filename. This isn't a perfect solution but it should do for now.