Automatically fill in user for API client shell
Use the new user helper to work out who the user is. This means that the user no longer needs to explicitly provide their user identifier on each API call that needs it.

Inform the user who they are in case this is wrong. The user can override the user for any API call if they wish.