Remove link to suppliers service
The URL we got was the link to the service in the API. This is not the same as the URL on the Digital Marketplace website, which is what I wanted.

I don't think there's a way to get the link I want, so just remove it and tell the person running the script what they need to do instead.