export-dos-opportunities: Add admin-only CSV with buyer details
Currently CSS admins have to cross-reference buyer user details with an
opportunity by hand. This commit expands the export-dos-opportunities
script to generate an additional CSV with extra details about the buyer
user. As we don't want to make this information publicly available, this
new CSV is uploaded to the reports bucket with private permissions; a
change to the admin frontend will be needed to allow admins to download
this file.