Get Trello cards with CSV attachments
Create a script to get us all our Trello cards with CSV attachments. This needs to be run by someone with access to all our Trello teams/boards. They'll need to get a Trello API key and token and provide them as environment variables.

We're doing this because we think some of the CSV attachments in Trello contain user data (emails, mostly). We want to find all such attachments so that we can delete them.