Change script that suspends suppliers to also send a reminder email
Previously this script suspended suppliers who have not signed the
framework agreement and exported a CSV of email addresses to notify.
These email addresses were then uploaded to Notify and a reminder
email was sent manually.

In the interests of saving time, reducing the scope for errors and
not having personal data on our laptops, modify the script so it
also sends the emails.

https://trello.com/c/ieH4CTed/699-1-make-script-to-suspend-suppliers-who-havent-signed-also-send-the-email